pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tim"
version = "0.1.0"
authors = [
    { name = "bpqoc", email="bpawlowski@qocinnovations.com" },
]
description = "Tool for time trackign"
readme = "README.md"
requires-python = ">=3.0"
dependencies = [
    "platformdirs",
    "textual",
    "python-dateutil",
]

[project.urls]
Homepage = "https://github.com/bpqoc/tim"
Issues = "https://github.com/bpqoc/tim/issues"

[project.scripts]
tim = "tim.main:main"
tim-gui = "tim.gui:main"